# serverless/JavaUnitTests/tagEngine/TagEngine-dev-swagger-postman.json
{
  "swagger": "2.0",
  "info": {
    "version": "2018-08-07T10:29:49Z",
    "title": "TagEngine"
  },
  "host": "ewqp8ot918.execute-api.us-east-1.amazonaws.com",
  "basePath": "/dev",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/disciplines/{discipline}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "discipline",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/Empty"
            }
          }
        }
      }
    },
...
    "/tag/{id}": {
      "put": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "$ref": "#/definitions/Empty"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Empty": {
      "type": "object",
      "title": "Empty Schema"
    }
  }
}